middleware/requirements_3_7.txt
aniso8601==4.0.1
asn1crypto==0.24.0
attrdict==2.0.0
base58==1.0.2
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
coincurve==11.0.0
colorama==0.4.1
coloredlogs==10.0
configparser==3.5.0
coverage==4.5.2
cytoolz==0.9.0.1
dictionaries==0.0.1
eth-abi==1.3.0
eth-account==0.3.0
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.2.0b3
eth-rlp==0.1.2
eth-tester==0.1.0b33
eth-typing==2.0.0
eth-utils==1.4.0
ethereum==2.3.2
ethereum-dasm==0.1.4
ethereum-input-decoder==0.2.2
evmdasm==0.1.8
Flask==1.0.2
flask-marshmallow==0.9.0
Flask-RESTful==0.3.7
future==0.17.1
hexbytes==0.1.0
humanfriendly==4.17
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
kafka-python==1.4.4
kazoo==2.5.0
lru-dict==1.1.6
MarkupSafe==1.1.0
marshmallow==2.17.0
mock==2.0.0
morphys==1.0
mythril==0.19.11
parsimonious==0.8.1
pbkdf2==1.3
pbr==5.1.1
persistent==4.4.3
plyvel==1.0.5
py-cid==0.2.1
py-ecc==1.4.2
py-flags==1.1.2
py-multibase==1.0.1
py-multicodec==0.1.3
py-multihash==0.2.3
py-solc==3.2.0
pycparser==2.19
pycryptodome==3.7.2
pyethash==0.1.27
pykafka==2.8.0
pysha3==1.0.2
python-baseconv==1.2.1
pytz==2018.7
PyYAML==3.13
redis==3.0.1
repoze.lru==0.7
requests==2.21.0
rlp==1.0.3
scrypt==0.8.6
semantic-version==2.6.0
six==1.11.0
solidity-flattener==0.2.2
tabulate==0.8.2
toolz==0.9.0
transaction==2.4.0
urllib3==1.24.1
varint==1.0.2
web3==4.8.2
websockets==6.0
zope.interface==4.6.0
